Bob/requirements.txt
django==1.9.1
djangorestframework==3.3.2
django-rest-swagger==0.3.5